feat(adm/4): chapter 4, its core exercises, and phase 2 completed

+++ study/adm/04-sorting/quicksort/quicksort.c
@@ -37,7 +37,7 @@ void print_arr(int *arr, size_t len) {
     printf("[");
     if (len)
         printf("%d", arr[0]);
-    for (int i = 1; i < len; i++)
+    for (size_t i = 1; i < len; i++)
         printf(", %d", arr[i]);
     printf("]\n");
 }
